DataModeling_Cassandra/requirements/requirements.txt
pandas
psycopg2
numpy
cassandra-driver